first iter of the tile manager
pretty simple, but good start, I think.
implemented a dumb hack to have the window not be resizable in my WM... TODO: check if this is necessary elsewhere or there is a cleaner way.
Deleted the aspect ratio correction from the orthoProjection, since 16:9 will be enforced (I really tried, but I have not found a satisfying way of allowing other aspect ratios)