minor fixes
Deserialization was being called 4 times on each keypress.
I have not researched this, but I am assuming it is because the keys
are polled at a pretty slow rate.
I don't really care about this (for now) because the memory should be
written and read just once, so I have mimicked that behaviour.

Also, instead of casting every single time the entity, I made it static
so it persists locally. I have yet to implement a decent system for entities
and components, but this seemed like an obvious and effortless improvement