major refactor of the input system
I am still not very happy about how this works, but this is way better;
Instead of using a map and polling it, just use SDL_GetKeyboardState and query that.
To handle "justPressed/Released" vs. heldDown, I created a new array for actions, that hold if the action has already been
triggered without a key release happening, and vivecersa.
The same is true for the mouse

This is way simpler and way cleaner (I just like it more tbh)
I still have to think about how to handle Joysticks (should not be too hard)