Basic & dumb layered rendering. Fixed bug. Playing around with Delaunator-cpp
Right now, it is basically just iterating again over all chunks just to
draw the different parts... I do not really want to rewrite the renderer
to handle z-ordering or whatever, so this will work for now.
Fixed a "bug" where the upmost tile was being loaded too late, causing
the player to see the clear color (basically, the window is actually
23.something tiles, not just 23)
Trying some things for the level generation.